main container-cards-profile
The container-card-profile is given a fax display  and inside it there are two div with main-box classes one and two, whose width is specified with   flex-grow to be responsive.
Main-box 1 and 2 have also been given flex display and flex warp is equal to warp so that the items are responsive.